[cryosat2 mds] changed output data types from float64 to float32 for most of the adapters

+++ pysiral/cryosat2/l1b_mds_def.py
@@ -8,30 +8,32 @@ Created on Fri Jul 10 21:00:37 2015
 from construct import (Struct, Array, Padding, Adapter, Bit, BitStruct,
                        SBInt16, UBInt16, SBInt32, UBInt32, SBInt64, UBInt64)
 
+import numpy as np
+
 
 class OneHundredth(Adapter):
     def _decode(self, obj, context):
-        return float(obj)/100
+        return np.float32(float(obj)/100)
 
 
 class OneHundredthDecibel(Adapter):
     def _decode(self, obj, context):
-        return float(obj)/100
+        return np.float32(float(obj)/100)
 
 
 class OneTenthMicroDeg(Adapter):
     def _decode(self, obj, context):
-        return float(obj)*1e-7
+        return np.float32(float(obj)*1e-7)
 
 
 class MilliMeter(Adapter):
     def _decode(self, obj, context):
-        return float(obj)*1e-3
+        return np.float32(float(obj)*1e-3)
 
 
 class Micrometer(Adapter):
     def _decode(self, obj, context):
-        return float(obj)*1e-6
+        return np.float32(float(obj)*1e-6)
 
 
 class PicoSecond(Adapter):
@@ -41,17 +43,17 @@ class PicoSecond(Adapter):
 
 class MicroWatts(Adapter):
     def _decode(self, obj, context):
-        return float(obj)*1e-6
+        return np.float32(float(obj)*1e-6)
 
 
 class MicroRadians(Adapter):
     def _decode(self, obj, context):
-        return float(obj)*1e-6
+        return np.float32(float(obj)*1e-6)
 
 
 class OneTenthMicroRadians(Adapter):
     def _decode(self, obj, context):
-        return float(obj)*1e-6
+        return np.float32(float(obj)*1e-6)
 
 
 class Cryosat2L1bMDSDefinition(object):
